added

+++ crud-with-table-relations/src/main/java/com/example/crud/with/table/relations/service/StudentService.java
@@ -1,6 +1,10 @@
 package com.example.crud.with.table.relations.service;
 
 import com.example.crud.with.table.relations.dto.StudentDto;
+import com.example.crud.with.table.relations.dto.TeacherDto;
+import com.example.crud.with.table.relations.entity.Student;
+import com.example.crud.with.table.relations.entity.Teacher;
+import com.example.crud.with.table.relations.exception.ProductNotFoundException;
 import com.example.crud.with.table.relations.repository.StudentRepo;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
@@ -10,6 +14,9 @@ import org.springframework.stereotype.Service;
 import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import java.util.stream.StreamSupport;
 
 @Service
 @Transactional
@@ -19,6 +26,7 @@ public class StudentService {
 
     @Autowired
     private ModelMapper modelMapper;
+
     // insert employee
     public void insertStudent(StudentDto studentDto)
     {
@@ -42,6 +50,7 @@ public class StudentService {
         return modelMapper.map(student, StudentDto.class);
     }
     public void deleteByStudentId(int studentId){
+        if(!studentRepo.existsById(studentId))throw new ProductNotFoundException();
         studentRepo.deleteById(studentId);
     }
 }
